requirements.txt
FLask==2.0.3
Pillow==10.4.0
pyzbar==0.1.9
zbarlight==3.0
